Started to implement entities for blog post implementations + new auth system

+++ src/main/java/Pradanoia/entities/UserInfo.java
@@ -1,14 +1,9 @@
 package Pradanoia.entities;
 
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
+import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-
-import java.util.Collection;
 
 @Entity
 @Data
@@ -17,7 +12,7 @@ import java.util.Collection;
 public class UserInfo {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
     private String name;
     private String email;
     private String password;
